Macros written for duplications, unit tests added, operations moved to kernel_arrow
