+++ datafusion/physical-expr/src/expressions/binary/kernels_arrow.rs
@@ -32,7 +32,7 @@ use arrow_array::types::{ArrowDictionaryKeyType, DecimalType};
 use arrow_array::ArrowNativeTypeOp;
 use arrow_buffer::ArrowNativeType;
 use arrow_schema::{DataType, IntervalUnit};
-use chrono::{Days, Duration, Months, NaiveDate};
+use chrono::{Days, Duration, Months, NaiveDate, NaiveDateTime};
 use datafusion_common::cast::{as_date32_array, as_date64_array, as_decimal128_array};
 use datafusion_common::scalar::{date32_op, date64_op};
 use datafusion_common::{DataFusionError, Result, ScalarValue};
@@ -40,13 +40,25 @@ use std::cmp::min;
 use std::ops::Add;
 use std::sync::Arc;
 
-use super::{
-    date32_interval_dt_op, date32_interval_mdn_op, date32_interval_ym_op,
-    date64_interval_dt_op, date64_interval_mdn_op, date64_interval_ym_op,
-    interval_array_op, interval_op_scalar_interval, scalar_interval_op_interval,
-    scalar_ts_op_interval, scalar_ts_sub_ts, ts_array_op, ts_interval_array_op,
-    ts_op_scalar_interval, ts_sub_scalar_ts,
+use arrow::compute::unary;
+use arrow::datatypes::*;
+
+use arrow_array::temporal_conversions::{MILLISECONDS_IN_DAY, NANOSECONDS_IN_DAY};
+use datafusion_common::delta::shift_months;
+use datafusion_common::scalar::{
+    calculate_naives, microseconds_add, microseconds_sub, milliseconds_add,
+    milliseconds_sub, nanoseconds_add, nanoseconds_sub, op_dt, op_dt_mdn, op_mdn, op_ym,
+    op_ym_dt, op_ym_mdn, parse_timezones, seconds_add, MILLISECOND_MODE, NANOSECOND_MODE,
 };
+
+use arrow::datatypes::TimeUnit;
+
+use datafusion_common::cast::{
+    as_interval_dt_array, as_interval_mdn_array, as_interval_ym_array,
+    as_timestamp_microsecond_array, as_timestamp_millisecond_array,
+    as_timestamp_nanosecond_array, as_timestamp_second_array,
+};
+use datafusion_common::scalar::*;
 
 // Simple (low performance) kernels until optimized kernels are added to arrow
 // See https://github.com/apache/arrow-rs/issues/960
@@ -871,6 +883,1297 @@ pub(crate) fn modulus_decimal_dyn_scalar(
 
     let array = modulus_scalar_dyn::<Decimal128Type>(left, right)?;
     decimal_array_with_precision_scale(array, precision, scale)
+}
+
+macro_rules! sub_timestamp_macro {
+    ($array:expr, $rhs:expr, $caster:expr, $interval_type:ty, $opt_tz_lhs:expr, $multiplier:expr,
+        $opt_tz_rhs:expr, $unit_sub:expr, $naive_sub_fn:expr, $counter:expr) => {{
+        let prim_array = $caster(&$array)?;
+        let ret: PrimitiveArray<$interval_type> = try_unary(prim_array, |lhs| {
+            let (parsed_lhs_tz, parsed_rhs_tz) =
+                (parse_timezones($opt_tz_lhs)?, parse_timezones($opt_tz_rhs)?);
+            let (naive_lhs, naive_rhs) = calculate_naives::<$unit_sub>(
+                lhs.mul_wrapping($multiplier),
+                parsed_lhs_tz,
+                $rhs.mul_wrapping($multiplier),
+                parsed_rhs_tz,
+            )?;
+            Ok($naive_sub_fn($counter(&naive_lhs), $counter(&naive_rhs)))
+        })?;
+        Arc::new(ret) as _
+    }};
+}
+
+macro_rules! sub_timestamp_left_scalar_macro {
+    ($array:expr, $lhs:expr, $caster:expr, $interval_type:ty, $opt_tz_lhs:expr, $multiplier:expr,
+        $opt_tz_rhs:expr, $unit_sub:expr, $naive_sub_fn:expr, $counter:expr) => {{
+        let prim_array = $caster(&$array)?;
+        let ret: PrimitiveArray<$interval_type> = try_unary(prim_array, |rhs| {
+            let (parsed_lhs_tz, parsed_rhs_tz) =
+                (parse_timezones($opt_tz_lhs)?, parse_timezones($opt_tz_rhs)?);
+            let (naive_lhs, naive_rhs) = calculate_naives::<$unit_sub>(
+                $lhs.mul_wrapping($multiplier),
+                parsed_lhs_tz,
+                rhs.mul_wrapping($multiplier),
+                parsed_rhs_tz,
+            )?;
+            Ok($naive_sub_fn($counter(&naive_lhs), $counter(&naive_rhs)))
+        })?;
+        Arc::new(ret) as _
+    }};
+}
+
+macro_rules! op_timestamp_interval_macro {
+    ($array:expr, $as_timestamp:expr, $ts_type:ty, $fn_op:expr, $scalar:expr, $sign:expr, $tz:expr) => {{
+        let array = $as_timestamp(&$array)?;
+        let ret: PrimitiveArray<$ts_type> =
+            try_unary::<$ts_type, _, $ts_type>(array, |ts_s| {
+                Ok($fn_op(ts_s, $scalar, $sign)?)
+            })?;
+        Arc::new(ret.with_timezone_opt($tz.clone())) as _
+    }};
+}
+
+macro_rules! scalar_ts_op_interval_macro {
+    ($ts:ident, $tz:ident, $interval:ident, $sign:ident,
+        $caster1:expr, $type1:ty, $type2:ty, $op:expr, $back_caster:expr) => {{
+        let interval = $caster1(&$interval)?;
+        let ret: PrimitiveArray<$type1> =
+            try_unary::<$type2, _, $type1>(interval, |e| {
+                let prior = $ts.ok_or_else(|| {
+                    DataFusionError::Internal("Timestamp is out-of-range".to_string())
+                })?;
+                Ok($back_caster(&$op(prior, e, $sign)))
+            })?;
+        Arc::new(ret.with_timezone_opt($tz.clone())) as _
+    }};
+}
+
+macro_rules! op_interval_macro {
+    ($array:expr, $as_interval:expr, $interval_type:ty, $fn_op:expr, $scalar:expr, $sign:expr) => {{
+        let array = $as_interval(&$array)?;
+        let ret: PrimitiveArray<$interval_type> =
+            unary(array, |lhs| $fn_op(lhs, *$scalar, $sign));
+        Arc::new(ret) as _
+    }};
+}
+
+macro_rules! op_interval_cross_macro {
+    ($array:expr, $as_interval:expr, $commute:expr, $fn_op:expr, $scalar:expr, $sign:expr, $t1:ty, $t2:ty) => {{
+        let array = $as_interval(&$array)?;
+        let ret: PrimitiveArray<IntervalMonthDayNanoType> = if $commute {
+            unary(array, |lhs| {
+                $fn_op(*$scalar as $t1, lhs as $t2, $sign, $commute)
+            })
+        } else {
+            unary(array, |lhs| {
+                $fn_op(lhs as $t1, *$scalar as $t2, $sign, $commute)
+            })
+        };
+        Arc::new(ret) as _
+    }};
+}
+
+macro_rules! ts_sub_op {
+    ($lhs:ident, $rhs:ident, $lhs_tz:ident, $rhs_tz:ident, $coef:expr, $caster:expr, $op:expr, $ts_unit:expr, $mode:expr, $type_out:ty) => {{
+        let prim_array_lhs = $caster(&$lhs)?;
+        let prim_array_rhs = $caster(&$rhs)?;
+        let ret: PrimitiveArray<$type_out> =
+            arrow::compute::try_binary(prim_array_lhs, prim_array_rhs, |ts1, ts2| {
+                let (parsed_lhs_tz, parsed_rhs_tz) = (
+                    parse_timezones($lhs_tz.as_deref())?,
+                    parse_timezones($rhs_tz.as_deref())?,
+                );
+                let (naive_lhs, naive_rhs) = calculate_naives::<$mode>(
+                    ts1.mul_wrapping($coef),
+                    parsed_lhs_tz,
+                    ts2.mul_wrapping($coef),
+                    parsed_rhs_tz,
+                )?;
+                Ok($op($ts_unit(&naive_lhs), $ts_unit(&naive_rhs)))
+            })?;
+        Arc::new(ret) as _
+    }};
+}
+
+macro_rules! interval_op {
+    ($lhs:ident, $rhs:ident, $caster:expr, $op:expr, $sign:ident, $type_in:ty) => {{
+        let prim_array_lhs = $caster(&$lhs)?;
+        let prim_array_rhs = $caster(&$rhs)?;
+        Arc::new(arrow::compute::binary::<$type_in, $type_in, _, $type_in>(
+            prim_array_lhs,
+            prim_array_rhs,
+            |interval1, interval2| $op(interval1, interval2, $sign),
+        )?) as _
+    }};
+}
+
+macro_rules! interval_cross_op {
+    ($lhs:ident, $rhs:ident, $caster1:expr, $caster2:expr, $op:expr, $sign:ident, $commute:ident, $type_in1:ty, $type_in2:ty) => {{
+        let prim_array_lhs = $caster1(&$lhs)?;
+        let prim_array_rhs = $caster2(&$rhs)?;
+        Arc::new(arrow::compute::binary::<
+            $type_in1,
+            $type_in2,
+            _,
+            IntervalMonthDayNanoType,
+        >(
+            prim_array_lhs,
+            prim_array_rhs,
+            |interval1, interval2| $op(interval1, interval2, $sign, $commute),
+        )?) as _
+    }};
+}
+
+macro_rules! ts_interval_op {
+    ($lhs:ident, $rhs:ident, $tz:ident, $caster1:expr, $caster2:expr, $op:expr, $sign:ident, $type_in1:ty, $type_in2:ty) => {{
+        let prim_array_lhs = $caster1(&$lhs)?;
+        let prim_array_rhs = $caster2(&$rhs)?;
+        let ret: PrimitiveArray<$type_in1> = arrow::compute::try_binary(
+            prim_array_lhs,
+            prim_array_rhs,
+            |ts, interval| Ok($op(ts, interval as i128, $sign)?),
+        )?;
+        Arc::new(ret.with_timezone_opt($tz.clone())) as _
+    }};
+}
+
+/// This function handles timestamp - timestamp operations where the former is
+/// an array and the latter is a scalar, resulting in an array.
+pub fn ts_sub_scalar_ts(array: &ArrayRef, scalar: &ScalarValue) -> Result<ArrayRef> {
+    let ret = match (array.data_type(), scalar) {
+        (
+            DataType::Timestamp(TimeUnit::Second, opt_tz_lhs),
+            ScalarValue::TimestampSecond(Some(rhs), opt_tz_rhs),
+        ) => {
+            sub_timestamp_macro!(
+                array,
+                rhs,
+                as_timestamp_second_array,
+                IntervalDayTimeType,
+                opt_tz_lhs.as_deref(),
+                1000,
+                opt_tz_rhs.as_deref(),
+                MILLISECOND_MODE,
+                seconds_sub,
+                NaiveDateTime::timestamp
+            )
+        }
+        (
+            DataType::Timestamp(TimeUnit::Millisecond, opt_tz_lhs),
+            ScalarValue::TimestampMillisecond(Some(rhs), opt_tz_rhs),
+        ) => {
+            sub_timestamp_macro!(
+                array,
+                rhs,
+                as_timestamp_millisecond_array,
+                IntervalDayTimeType,
+                opt_tz_lhs.as_deref(),
+                1,
+                opt_tz_rhs.as_deref(),
+                MILLISECOND_MODE,
+                milliseconds_sub,
+                NaiveDateTime::timestamp_millis
+            )
+        }
+        (
+            DataType::Timestamp(TimeUnit::Microsecond, opt_tz_lhs),
+            ScalarValue::TimestampMicrosecond(Some(rhs), opt_tz_rhs),
+        ) => {
+            sub_timestamp_macro!(
+                array,
+                rhs,
+                as_timestamp_microsecond_array,
+                IntervalMonthDayNanoType,
+                opt_tz_lhs.as_deref(),
+                1000,
+                opt_tz_rhs.as_deref(),
+                NANOSECOND_MODE,
+                microseconds_sub,
+                NaiveDateTime::timestamp_micros
+            )
+        }
+        (
+            DataType::Timestamp(TimeUnit::Nanosecond, opt_tz_lhs),
+            ScalarValue::TimestampNanosecond(Some(rhs), opt_tz_rhs),
+        ) => {
+            sub_timestamp_macro!(
+                array,
+                rhs,
+                as_timestamp_nanosecond_array,
+                IntervalMonthDayNanoType,
+                opt_tz_lhs.as_deref(),
+                1,
+                opt_tz_rhs.as_deref(),
+                NANOSECOND_MODE,
+                nanoseconds_sub,
+                NaiveDateTime::timestamp_nanos
+            )
+        }
+        (_, _) => {
+            return Err(DataFusionError::Internal(format!(
+                "Invalid array - scalar types for Timestamp subtraction: {:?} - {:?}",
+                array.data_type(),
+                scalar.get_datatype()
+            )));
+        }
+    };
+    Ok(ret)
+}
+
+/// This function handles timestamp - timestamp operations where the former is
+/// a scalar and the latter is an array, resulting in an array.
+pub fn scalar_ts_sub_ts(scalar: &ScalarValue, array: &ArrayRef) -> Result<ArrayRef> {
+    let ret = match (scalar, array.data_type()) {
+        (
+            ScalarValue::TimestampSecond(Some(lhs), opt_tz_lhs),
+            DataType::Timestamp(TimeUnit::Second, opt_tz_rhs),
+        ) => {
+            sub_timestamp_left_scalar_macro!(
+                array,
+                lhs,
+                as_timestamp_second_array,
+                IntervalDayTimeType,
+                opt_tz_lhs.as_deref(),
+                1000,
+                opt_tz_rhs.as_deref(),
+                MILLISECOND_MODE,
+                seconds_sub,
+                NaiveDateTime::timestamp
+            )
+        }
+        (
+            ScalarValue::TimestampMillisecond(Some(lhs), opt_tz_lhs),
+            DataType::Timestamp(TimeUnit::Millisecond, opt_tz_rhs),
+        ) => {
+            sub_timestamp_left_scalar_macro!(
+                array,
+                lhs,
+                as_timestamp_millisecond_array,
+                IntervalDayTimeType,
+                opt_tz_lhs.as_deref(),
+                1,
+                opt_tz_rhs.as_deref(),
+                MILLISECOND_MODE,
+                milliseconds_sub,
+                NaiveDateTime::timestamp_millis
+            )
+        }
+        (
+            ScalarValue::TimestampMicrosecond(Some(lhs), opt_tz_lhs),
+            DataType::Timestamp(TimeUnit::Microsecond, opt_tz_rhs),
+        ) => {
+            sub_timestamp_left_scalar_macro!(
+                array,
+                lhs,
+                as_timestamp_microsecond_array,
+                IntervalMonthDayNanoType,
+                opt_tz_lhs.as_deref(),
+                1000,
+                opt_tz_rhs.as_deref(),
+                NANOSECOND_MODE,
+                microseconds_sub,
+                NaiveDateTime::timestamp_micros
+            )
+        }
+        (
+            ScalarValue::TimestampNanosecond(Some(lhs), opt_tz_lhs),
+            DataType::Timestamp(TimeUnit::Nanosecond, opt_tz_rhs),
+        ) => {
+            sub_timestamp_left_scalar_macro!(
+                array,
+                lhs,
+                as_timestamp_nanosecond_array,
+                IntervalMonthDayNanoType,
+                opt_tz_lhs.as_deref(),
+                1,
+                opt_tz_rhs.as_deref(),
+                NANOSECOND_MODE,
+                nanoseconds_sub,
+                NaiveDateTime::timestamp_nanos
+            )
+        }
+        (_, _) => {
+            return Err(DataFusionError::Internal(format!(
+                "Invalid scalar - array types for Timestamp subtraction: {:?} - {:?}",
+                scalar.get_datatype(),
+                array.data_type()
+            )));
+        }
+    };
+    Ok(ret)
+}
+
+/// This function handles timestamp +/- interval operations where the former is
+/// an array and the latter is a scalar, resulting in an array.
+pub fn ts_op_scalar_interval(
+    array: &ArrayRef,
+    sign: i32,
+    scalar: &ScalarValue,
+) -> Result<ArrayRef> {
+    let ret = match array.data_type() {
+        DataType::Timestamp(TimeUnit::Second, tz) => {
+            op_timestamp_interval_macro!(
+                array,
+                as_timestamp_second_array,
+                TimestampSecondType,
+                seconds_add,
+                scalar,
+                sign,
+                tz
+            )
+        }
+        DataType::Timestamp(TimeUnit::Millisecond, tz) => {
+            op_timestamp_interval_macro!(
+                array,
+                as_timestamp_millisecond_array,
+                TimestampMillisecondType,
+                milliseconds_add,
+                scalar,
+                sign,
+                tz
+            )
+        }
+        DataType::Timestamp(TimeUnit::Microsecond, tz) => {
+            op_timestamp_interval_macro!(
+                array,
+                as_timestamp_microsecond_array,
+                TimestampMicrosecondType,
+                microseconds_add,
+                scalar,
+                sign,
+                tz
+            )
+        }
+        DataType::Timestamp(TimeUnit::Nanosecond, tz) => {
+            op_timestamp_interval_macro!(
+                array,
+                as_timestamp_nanosecond_array,
+                TimestampNanosecondType,
+                nanoseconds_add,
+                scalar,
+                sign,
+                tz
+            )
+        }
+        _ => Err(DataFusionError::Internal(format!(
+            "Invalid lhs type for Timestamp vs Interval operations: {}",
+            array.data_type()
+        )))?,
+    };
+    Ok(ret)
+}
+
+/// This function handles timestamp +/- interval operations where the former is
+/// a scalar and the latter is an array, resulting in an array.
+pub fn scalar_ts_op_interval(
+    scalar: &ScalarValue,
+    sign: i32,
+    array: &ArrayRef,
+) -> Result<ArrayRef> {
+    use DataType::*;
+    use IntervalUnit::*;
+    use ScalarValue::*;
+    let ret = match (scalar, array.data_type()) {
+        // Second op YearMonth
+        (TimestampSecond(Some(ts_sec), tz), Interval(YearMonth)) => {
+            let naive_date = NaiveDateTime::from_timestamp_opt(*ts_sec, 0);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_ym_array,
+                TimestampSecondType,
+                IntervalYearMonthType,
+                shift_months,
+                NaiveDateTime::timestamp
+            )
+        }
+        // Millisecond op YearMonth
+        (TimestampMillisecond(Some(ts_ms), tz), Interval(YearMonth)) => {
+            let naive_date = NaiveDateTime::from_timestamp_millis(*ts_ms);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_ym_array,
+                TimestampSecondType,
+                IntervalYearMonthType,
+                shift_months,
+                NaiveDateTime::timestamp
+            )
+        }
+        // Microsecond op YearMonth
+        (TimestampMicrosecond(Some(ts_us), tz), Interval(YearMonth)) => {
+            let naive_date = NaiveDateTime::from_timestamp_micros(*ts_us);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_ym_array,
+                TimestampSecondType,
+                IntervalYearMonthType,
+                shift_months,
+                NaiveDateTime::timestamp
+            )
+        }
+        // Nanosecond op YearMonth
+        (TimestampNanosecond(Some(ts_ns), tz), Interval(YearMonth)) => {
+            let naive_date = NaiveDateTime::from_timestamp_opt(
+                ts_ns.div_euclid(1_000_000_000),
+                ts_ns.rem_euclid(1_000_000_000).try_into().map_err(|_| {
+                    DataFusionError::Internal("Overflow of divison".to_string())
+                })?,
+            );
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_ym_array,
+                TimestampSecondType,
+                IntervalYearMonthType,
+                shift_months,
+                NaiveDateTime::timestamp
+            )
+        }
+        // Second op DayTime
+        (TimestampSecond(Some(ts_sec), tz), Interval(DayTime)) => {
+            let naive_date = NaiveDateTime::from_timestamp_opt(*ts_sec, 0);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_dt_array,
+                TimestampSecondType,
+                IntervalDayTimeType,
+                add_day_time,
+                NaiveDateTime::timestamp
+            )
+        }
+        // Millisecond op DayTime
+        (TimestampMillisecond(Some(ts_ms), tz), Interval(DayTime)) => {
+            let naive_date = NaiveDateTime::from_timestamp_millis(*ts_ms);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_dt_array,
+                TimestampMillisecondType,
+                IntervalDayTimeType,
+                add_day_time,
+                NaiveDateTime::timestamp_millis
+            )
+        }
+        // Microsecond op DayTime
+        (TimestampMicrosecond(Some(ts_us), tz), Interval(DayTime)) => {
+            let naive_date = NaiveDateTime::from_timestamp_micros(*ts_us);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_dt_array,
+                TimestampMicrosecondType,
+                IntervalDayTimeType,
+                add_day_time,
+                NaiveDateTime::timestamp_micros
+            )
+        }
+        // Nanosecond op DayTime
+        (TimestampNanosecond(Some(ts_ns), tz), Interval(DayTime)) => {
+            let naive_date = NaiveDateTime::from_timestamp_opt(
+                ts_ns.div_euclid(1_000_000_000),
+                ts_ns.rem_euclid(1_000_000_000).try_into().map_err(|_| {
+                    DataFusionError::Internal("Overflow of divison".to_string())
+                })?,
+            );
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_dt_array,
+                TimestampNanosecondType,
+                IntervalDayTimeType,
+                add_day_time,
+                NaiveDateTime::timestamp_nanos
+            )
+        }
+        // Second op MonthDayNano
+        (TimestampSecond(Some(ts_sec), tz), Interval(MonthDayNano)) => {
+            let naive_date = NaiveDateTime::from_timestamp_opt(*ts_sec, 0);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_mdn_array,
+                TimestampSecondType,
+                IntervalMonthDayNanoType,
+                add_m_d_nano,
+                NaiveDateTime::timestamp
+            )
+        }
+        // Millisecond op MonthDayNano
+        (TimestampMillisecond(Some(ts_ms), tz), Interval(MonthDayNano)) => {
+            let naive_date = NaiveDateTime::from_timestamp_millis(*ts_ms);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_mdn_array,
+                TimestampMillisecondType,
+                IntervalMonthDayNanoType,
+                add_m_d_nano,
+                NaiveDateTime::timestamp_millis
+            )
+        }
+        // Microsecond op MonthDayNano
+        (TimestampMicrosecond(Some(ts_us), tz), Interval(MonthDayNano)) => {
+            let naive_date = NaiveDateTime::from_timestamp_micros(*ts_us);
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_mdn_array,
+                TimestampMicrosecondType,
+                IntervalMonthDayNanoType,
+                add_m_d_nano,
+                NaiveDateTime::timestamp_micros
+            )
+        }
+
+        // Nanosecond op MonthDayNano
+        (TimestampNanosecond(Some(ts_ns), tz), Interval(MonthDayNano)) => {
+            let naive_date = NaiveDateTime::from_timestamp_opt(
+                ts_ns.div_euclid(1_000_000_000),
+                ts_ns.rem_euclid(1_000_000_000).try_into().map_err(|_| {
+                    DataFusionError::Internal("Overflow of divison".to_string())
+                })?,
+            );
+            scalar_ts_op_interval_macro!(
+                naive_date,
+                tz,
+                array,
+                sign,
+                as_interval_mdn_array,
+                TimestampNanosecondType,
+                IntervalMonthDayNanoType,
+                add_m_d_nano,
+                NaiveDateTime::timestamp_nanos
+            )
+        }
+        _ => Err(DataFusionError::Internal(
+            "Invalid types for Timestamp vs Interval operations".to_string(),
+        ))?,
+    };
+    Ok(ret)
+}
+
+/// This function handles interval +/- interval operations where the former is
+/// an array and the latter is a scalar, resulting in an interval array.
+pub fn interval_op_scalar_interval(
+    array: &ArrayRef,
+    sign: i32,
+    scalar: &ScalarValue,
+) -> Result<ArrayRef> {
+    use DataType::*;
+    use IntervalUnit::*;
+    use ScalarValue::*;
+    let ret = match (array.data_type(), scalar) {
+        (Interval(YearMonth), IntervalYearMonth(Some(rhs))) => {
+            op_interval_macro!(
+                array,
+                as_interval_ym_array,
+                IntervalYearMonthType,
+                op_ym,
+                rhs,
+                sign
+            )
+        }
+        (Interval(YearMonth), IntervalDayTime(Some(rhs))) => {
+            op_interval_cross_macro!(
+                array,
+                as_interval_ym_array,
+                false,
+                op_ym_dt,
+                rhs,
+                sign,
+                i32,
+                i64
+            )
+        }
+        (Interval(YearMonth), IntervalMonthDayNano(Some(rhs))) => {
+            op_interval_cross_macro!(
+                array,
+                as_interval_ym_array,
+                false,
+                op_ym_mdn,
+                rhs,
+                sign,
+                i32,
+                i128
+            )
+        }
+        (Interval(DayTime), IntervalYearMonth(Some(rhs))) => {
+            op_interval_cross_macro!(
+                array,
+                as_interval_dt_array,
+                true,
+                op_ym_dt,
+                rhs,
+                sign,
+                i32,
+                i64
+            )
+        }
+        (Interval(DayTime), IntervalDayTime(Some(rhs))) => {
+            op_interval_macro!(
+                array,
+                as_interval_dt_array,
+                IntervalDayTimeType,
+                op_dt,
+                rhs,
+                sign
+            )
+        }
+        (Interval(DayTime), IntervalMonthDayNano(Some(rhs))) => {
+            op_interval_cross_macro!(
+                array,
+                as_interval_dt_array,
+                false,
+                op_dt_mdn,
+                rhs,
+                sign,
+                i64,
+                i128
+            )
+        }
+        (Interval(MonthDayNano), IntervalYearMonth(Some(rhs))) => {
+            op_interval_cross_macro!(
+                array,
+                as_interval_mdn_array,
+                true,
+                op_ym_mdn,
+                rhs,
+                sign,
+                i32,
+                i128
+            )
+        }
+        (Interval(MonthDayNano), IntervalDayTime(Some(rhs))) => {
+            op_interval_cross_macro!(
+                array,
+                as_interval_mdn_array,
+                true,
+                op_dt_mdn,
+                rhs,
+                sign,
+                i64,
+                i128
+            )
+        }
+        (Interval(MonthDayNano), IntervalMonthDayNano(Some(rhs))) => {
+            op_interval_macro!(
+                array,
+                as_interval_mdn_array,
+                IntervalMonthDayNanoType,
+                op_mdn,
+                rhs,
+                sign
+            )
+        }
+        _ => Err(DataFusionError::Internal(format!(
+            "Invalid operands for Interval vs Interval operations: {} - {}",
+            array.data_type(),
+            scalar.get_datatype(),
+        )))?,
+    };
+    Ok(ret)
+}
+
+/// This function handles interval +/- interval operations where the former is
+/// a scalar and the latter is an array, resulting in an interval array.
+pub fn scalar_interval_op_interval(
+    scalar: &ScalarValue,
+    sign: i32,
+    array: &ArrayRef,
+) -> Result<ArrayRef> {
+    use DataType::*;
+    use IntervalUnit::*;
+    use ScalarValue::*;
+    let ret = match (scalar, array.data_type()) {
+        // YearMonth op YearMonth
+        (IntervalYearMonth(Some(lhs)), Interval(YearMonth)) => {
+            let array = as_interval_ym_array(&array)?;
+            let ret: PrimitiveArray<IntervalYearMonthType> =
+                unary(array, |rhs| op_ym(*lhs, rhs, sign));
+            Arc::new(ret) as _
+        }
+        // DayTime op YearMonth
+        (IntervalDayTime(Some(lhs)), Interval(YearMonth)) => {
+            let array = as_interval_ym_array(&array)?;
+            let ret: PrimitiveArray<IntervalMonthDayNanoType> =
+                unary(array, |rhs| op_ym_dt(rhs, *lhs, sign, true));
+            Arc::new(ret) as _
+        }
+        // MonthDayNano op YearMonth
+        (IntervalMonthDayNano(Some(lhs)), Interval(YearMonth)) => {
+            let array = as_interval_ym_array(&array)?;
+            let ret: PrimitiveArray<IntervalMonthDayNanoType> =
+                unary(array, |rhs| op_ym_mdn(rhs, *lhs, sign, true));
+            Arc::new(ret) as _
+        }
+        // YearMonth op DayTime
+        (IntervalYearMonth(Some(lhs)), Interval(DayTime)) => {
+            let array = as_interval_dt_array(&array)?;
+            let ret: PrimitiveArray<IntervalMonthDayNanoType> =
+                unary(array, |rhs| op_ym_dt(*lhs, rhs, sign, false));
+            Arc::new(ret) as _
+        }
+        // DayTime op DayTime
+        (IntervalDayTime(Some(lhs)), Interval(DayTime)) => {
+            let array = as_interval_dt_array(&array)?;
+            let ret: PrimitiveArray<IntervalDayTimeType> =
+                unary(array, |rhs| op_dt(*lhs, rhs, sign));
+            Arc::new(ret) as _
+        }
+        // MonthDayNano op DayTime
+        (IntervalMonthDayNano(Some(lhs)), Interval(DayTime)) => {
+            let array = as_interval_dt_array(&array)?;
+            let ret: PrimitiveArray<IntervalMonthDayNanoType> =
+                unary(array, |rhs| op_dt_mdn(rhs, *lhs, sign, true));
+            Arc::new(ret) as _
+        }
+        // YearMonth op MonthDayNano
+        (IntervalYearMonth(Some(lhs)), Interval(MonthDayNano)) => {
+            let array = as_interval_mdn_array(&array)?;
+            let ret: PrimitiveArray<IntervalMonthDayNanoType> =
+                unary(array, |rhs| op_ym_mdn(*lhs, rhs, sign, false));
+            Arc::new(ret) as _
+        }
+        // DayTime op MonthDayNano
+        (IntervalDayTime(Some(lhs)), Interval(MonthDayNano)) => {
+            let array = as_interval_mdn_array(&array)?;
+            let ret: PrimitiveArray<IntervalMonthDayNanoType> =
+                unary(array, |rhs| op_dt_mdn(*lhs, rhs, sign, false));
+            Arc::new(ret) as _
+        }
+        // MonthDayNano op MonthDayNano
+        (IntervalMonthDayNano(Some(lhs)), Interval(MonthDayNano)) => {
+            let array = as_interval_mdn_array(&array)?;
+            let ret: PrimitiveArray<IntervalMonthDayNanoType> =
+                unary(array, |rhs| op_mdn(*lhs, rhs, sign));
+            Arc::new(ret) as _
+        }
+        _ => Err(DataFusionError::Internal(format!(
+            "Invalid operands for Interval vs Interval operations: {} - {}",
+            scalar.get_datatype(),
+            array.data_type(),
+        )))?,
+    };
+    Ok(ret)
+}
+
+/// Performs a timestamp subtraction operation on two arrays and returns the resulting array.
+pub fn ts_array_op(array_lhs: &ArrayRef, array_rhs: &ArrayRef) -> Result<ArrayRef> {
+    use DataType::*;
+    use TimeUnit::*;
+    match (array_lhs.data_type(), array_rhs.data_type()) {
+        (Timestamp(Second, opt_tz_lhs), Timestamp(Second, opt_tz_rhs)) => Ok(ts_sub_op!(
+            array_lhs,
+            array_rhs,
+            opt_tz_lhs,
+            opt_tz_rhs,
+            1000i64,
+            as_timestamp_second_array,
+            seconds_sub,
+            NaiveDateTime::timestamp,
+            MILLISECOND_MODE,
+            IntervalDayTimeType
+        )),
+        (Timestamp(Millisecond, opt_tz_lhs), Timestamp(Millisecond, opt_tz_rhs)) => {
+            Ok(ts_sub_op!(
+                array_lhs,
+                array_rhs,
+                opt_tz_lhs,
+                opt_tz_rhs,
+                1i64,
+                as_timestamp_millisecond_array,
+                milliseconds_sub,
+                NaiveDateTime::timestamp_millis,
+                MILLISECOND_MODE,
+                IntervalDayTimeType
+            ))
+        }
+        (Timestamp(Microsecond, opt_tz_lhs), Timestamp(Microsecond, opt_tz_rhs)) => {
+            Ok(ts_sub_op!(
+                array_lhs,
+                array_rhs,
+                opt_tz_lhs,
+                opt_tz_rhs,
+                1000i64,
+                as_timestamp_microsecond_array,
+                microseconds_sub,
+                NaiveDateTime::timestamp_micros,
+                NANOSECOND_MODE,
+                IntervalMonthDayNanoType
+            ))
+        }
+        (Timestamp(Nanosecond, opt_tz_lhs), Timestamp(Nanosecond, opt_tz_rhs)) => {
+            Ok(ts_sub_op!(
+                array_lhs,
+                array_rhs,
+                opt_tz_lhs,
+                opt_tz_rhs,
+                1i64,
+                as_timestamp_nanosecond_array,
+                nanoseconds_sub,
+                NaiveDateTime::timestamp_nanos,
+                NANOSECOND_MODE,
+                IntervalMonthDayNanoType
+            ))
+        }
+        (_, _) => Err(DataFusionError::Execution(format!(
+            "Invalid array types for Timestamp subtraction: {} - {}",
+            array_lhs.data_type(),
+            array_rhs.data_type()
+        ))),
+    }
+}
+/// Performs an interval operation on two arrays and returns the resulting array.
+/// The operation sign determines whether to perform addition or subtraction.
+/// The data type and unit of the two input arrays must match the supported combinations.
+pub fn interval_array_op(
+    array_lhs: &ArrayRef,
+    array_rhs: &ArrayRef,
+    sign: i32,
+) -> Result<ArrayRef> {
+    use DataType::*;
+    use IntervalUnit::*;
+    match (array_lhs.data_type(), array_rhs.data_type()) {
+        (Interval(YearMonth), Interval(YearMonth)) => Ok(interval_op!(
+            array_lhs,
+            array_rhs,
+            as_interval_ym_array,
+            op_ym,
+            sign,
+            IntervalYearMonthType
+        )),
+        (Interval(YearMonth), Interval(DayTime)) => Ok(interval_cross_op!(
+            array_lhs,
+            array_rhs,
+            as_interval_ym_array,
+            as_interval_dt_array,
+            op_ym_dt,
+            sign,
+            false,
+            IntervalYearMonthType,
+            IntervalDayTimeType
+        )),
+        (Interval(YearMonth), Interval(MonthDayNano)) => Ok(interval_cross_op!(
+            array_lhs,
+            array_rhs,
+            as_interval_ym_array,
+            as_interval_mdn_array,
+            op_ym_mdn,
+            sign,
+            false,
+            IntervalYearMonthType,
+            IntervalMonthDayNanoType
+        )),
+        (Interval(DayTime), Interval(YearMonth)) => Ok(interval_cross_op!(
+            array_rhs,
+            array_lhs,
+            as_interval_ym_array,
+            as_interval_dt_array,
+            op_ym_dt,
+            sign,
+            true,
+            IntervalYearMonthType,
+            IntervalDayTimeType
+        )),
+        (Interval(DayTime), Interval(DayTime)) => Ok(interval_op!(
+            array_lhs,
+            array_rhs,
+            as_interval_dt_array,
+            op_dt,
+            sign,
+            IntervalDayTimeType
+        )),
+        (Interval(DayTime), Interval(MonthDayNano)) => Ok(interval_cross_op!(
+            array_lhs,
+            array_rhs,
+            as_interval_dt_array,
+            as_interval_mdn_array,
+            op_dt_mdn,
+            sign,
+            false,
+            IntervalDayTimeType,
+            IntervalMonthDayNanoType
+        )),
+        (Interval(MonthDayNano), Interval(YearMonth)) => Ok(interval_cross_op!(
+            array_rhs,
+            array_lhs,
+            as_interval_ym_array,
+            as_interval_mdn_array,
+            op_ym_mdn,
+            sign,
+            true,
+            IntervalYearMonthType,
+            IntervalMonthDayNanoType
+        )),
+        (Interval(MonthDayNano), Interval(DayTime)) => Ok(interval_cross_op!(
+            array_rhs,
+            array_lhs,
+            as_interval_dt_array,
+            as_interval_mdn_array,
+            op_dt_mdn,
+            sign,
+            true,
+            IntervalDayTimeType,
+            IntervalMonthDayNanoType
+        )),
+        (Interval(MonthDayNano), Interval(MonthDayNano)) => Ok(interval_op!(
+            array_lhs,
+            array_rhs,
+            as_interval_mdn_array,
+            op_mdn,
+            sign,
+            IntervalMonthDayNanoType
+        )),
+        (_, _) => Err(DataFusionError::Execution(format!(
+            "Invalid array types for Interval operation: {} {} {}",
+            array_lhs.data_type(),
+            sign,
+            array_rhs.data_type()
+        ))),
+    }
+}
+
+/// Performs a timestamp/interval operation on two arrays and returns the resulting array.
+/// The operation sign determines whether to perform addition or subtraction.
+/// The data type and unit of the two input arrays must match the supported combinations.
+pub fn ts_interval_array_op(
+    array_lhs: &ArrayRef,
+    sign: i32,
+    array_rhs: &ArrayRef,
+) -> Result<ArrayRef> {
+    use DataType::*;
+    use IntervalUnit::*;
+    use TimeUnit::*;
+    match (array_lhs.data_type(), array_rhs.data_type()) {
+        (Timestamp(Second, tz), Interval(YearMonth)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_second_array,
+            as_interval_ym_array,
+            seconds_add_array::<YM_MODE>,
+            sign,
+            TimestampSecondType,
+            IntervalYearMonthType
+        )),
+        (Timestamp(Second, tz), Interval(DayTime)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_second_array,
+            as_interval_dt_array,
+            seconds_add_array::<DT_MODE>,
+            sign,
+            TimestampSecondType,
+            IntervalDayTimeType
+        )),
+        (Timestamp(Second, tz), Interval(MonthDayNano)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_second_array,
+            as_interval_mdn_array,
+            seconds_add_array::<MDN_MODE>,
+            sign,
+            TimestampSecondType,
+            IntervalMonthDayNanoType
+        )),
+        (Timestamp(Millisecond, tz), Interval(YearMonth)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_millisecond_array,
+            as_interval_ym_array,
+            milliseconds_add_array::<YM_MODE>,
+            sign,
+            TimestampMillisecondType,
+            IntervalYearMonthType
+        )),
+        (Timestamp(Millisecond, tz), Interval(DayTime)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_millisecond_array,
+            as_interval_dt_array,
+            milliseconds_add_array::<DT_MODE>,
+            sign,
+            TimestampMillisecondType,
+            IntervalDayTimeType
+        )),
+        (Timestamp(Millisecond, tz), Interval(MonthDayNano)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_millisecond_array,
+            as_interval_mdn_array,
+            milliseconds_add_array::<MDN_MODE>,
+            sign,
+            TimestampMillisecondType,
+            IntervalMonthDayNanoType
+        )),
+        (Timestamp(Microsecond, tz), Interval(YearMonth)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_microsecond_array,
+            as_interval_ym_array,
+            microseconds_add_array::<YM_MODE>,
+            sign,
+            TimestampMicrosecondType,
+            IntervalYearMonthType
+        )),
+        (Timestamp(Microsecond, tz), Interval(DayTime)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_microsecond_array,
+            as_interval_dt_array,
+            microseconds_add_array::<DT_MODE>,
+            sign,
+            TimestampMicrosecondType,
+            IntervalDayTimeType
+        )),
+        (Timestamp(Microsecond, tz), Interval(MonthDayNano)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_microsecond_array,
+            as_interval_mdn_array,
+            microseconds_add_array::<MDN_MODE>,
+            sign,
+            TimestampMicrosecondType,
+            IntervalMonthDayNanoType
+        )),
+        (Timestamp(Nanosecond, tz), Interval(YearMonth)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_nanosecond_array,
+            as_interval_ym_array,
+            nanoseconds_add_array::<YM_MODE>,
+            sign,
+            TimestampNanosecondType,
+            IntervalYearMonthType
+        )),
+        (Timestamp(Nanosecond, tz), Interval(DayTime)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_nanosecond_array,
+            as_interval_dt_array,
+            nanoseconds_add_array::<DT_MODE>,
+            sign,
+            TimestampNanosecondType,
+            IntervalDayTimeType
+        )),
+        (Timestamp(Nanosecond, tz), Interval(MonthDayNano)) => Ok(ts_interval_op!(
+            array_lhs,
+            array_rhs,
+            tz,
+            as_timestamp_nanosecond_array,
+            as_interval_mdn_array,
+            nanoseconds_add_array::<MDN_MODE>,
+            sign,
+            TimestampNanosecondType,
+            IntervalMonthDayNanoType
+        )),
+        (_, _) => Err(DataFusionError::Execution(format!(
+            "Invalid array types for Timestamp Interval operation: {} {} {}",
+            array_lhs.data_type(),
+            sign,
+            array_rhs.data_type()
+        ))),
+    }
+}
+
+#[inline]
+pub fn date32_interval_ym_op(
+    right: &Arc<dyn Array>,
+    epoch: &NaiveDate,
+    prior: &NaiveDate,
+    month_op: fn(NaiveDate, Months) -> Option<NaiveDate>,
+) -> Result<ArrayRef> {
+    let right: &PrimitiveArray<IntervalYearMonthType> = right.as_primitive();
+    let ret = Arc::new(try_unary::<IntervalYearMonthType, _, Date32Type>(
+        right,
+        |ym| {
+            let months = Months::new(ym.try_into().map_err(|_| {
+                DataFusionError::Internal(
+                    "Interval values cannot be casted as unsigned integers".to_string(),
+                )
+            })?);
+            let value = month_op(*prior, months).ok_or_else(|| {
+                DataFusionError::Internal("Resulting date is out of range".to_string())
+            })?;
+            Ok((value - *epoch).num_days() as i32)
+        },
+    )?) as _;
+    Ok(ret)
+}
+
+#[inline]
+pub fn date32_interval_dt_op(
+    right: &Arc<dyn Array>,
+    epoch: &NaiveDate,
+    prior: &NaiveDate,
+    day_op: fn(NaiveDate, Days) -> Option<NaiveDate>,
+) -> Result<ArrayRef> {
+    let right: &PrimitiveArray<IntervalDayTimeType> = right.as_primitive();
+    let ret = Arc::new(try_unary::<IntervalDayTimeType, _, Date32Type>(
+        right,
+        |dt| {
+            let (days, millis) = IntervalDayTimeType::to_parts(dt);
+            let days = Days::new(days.try_into().map_err(|_| {
+                DataFusionError::Internal(
+                    "Interval values cannot be casted as unsigned integers".to_string(),
+                )
+            })?);
+            let value = day_op(*prior, days).ok_or_else(|| {
+                DataFusionError::Internal("Resulting date is out of range".to_string())
+            })?;
+            let milli_days = millis as i64 / MILLISECONDS_IN_DAY;
+            Ok(((value - *epoch).num_days() - milli_days) as i32)
+        },
+    )?) as _;
+    Ok(ret)
+}
+
+#[inline]
+pub fn date32_interval_mdn_op(
+    right: &Arc<dyn Array>,
+    epoch: &NaiveDate,
+    prior: &NaiveDate,
+    day_op: fn(NaiveDate, Days) -> Option<NaiveDate>,
+    month_op: fn(NaiveDate, Months) -> Option<NaiveDate>,
+) -> Result<ArrayRef> {
+    let cast_err = |_| {
+        DataFusionError::Internal(
+            "Interval values cannot be casted as unsigned integers".to_string(),
+        )
+    };
+    let out_of_range =
+        || DataFusionError::Internal("Resulting date is out of range".to_string());
+    let right: &PrimitiveArray<IntervalMonthDayNanoType> = right.as_primitive();
+    let ret = Arc::new(try_unary::<IntervalMonthDayNanoType, _, Date32Type>(
+        right,
+        |mdn| {
+            let (months, days, nanos) = IntervalMonthDayNanoType::to_parts(mdn);
+            let months_obj = Months::new(months.try_into().map_err(cast_err)?);
+            let month_diff = month_op(*prior, months_obj).ok_or_else(out_of_range)?;
+            let days_obj = Days::new(days.try_into().map_err(cast_err)?);
+            let value = day_op(month_diff, days_obj).ok_or_else(out_of_range)?;
+            let nano_days = nanos / NANOSECONDS_IN_DAY;
+            Ok(((value - *epoch).num_days() - nano_days) as i32)
+        },
+    )?) as _;
+    Ok(ret)
+}
+
+#[inline]
+pub fn date64_interval_ym_op(
+    right: &Arc<dyn Array>,
+    epoch: &NaiveDate,
+    prior: &NaiveDate,
+    month_op: fn(NaiveDate, Months) -> Option<NaiveDate>,
+) -> Result<ArrayRef> {
+    let right: &PrimitiveArray<IntervalYearMonthType> = right.as_primitive();
+    let ret = Arc::new(try_unary::<IntervalYearMonthType, _, Date64Type>(
+        right,
+        |ym| {
+            let months_obj = Months::new(ym.try_into().map_err(|_| {
+                DataFusionError::Internal(
+                    "Interval values cannot be casted as unsigned integers".to_string(),
+                )
+            })?);
+            let date = month_op(*prior, months_obj).ok_or_else(|| {
+                DataFusionError::Internal("Resulting date is out of range".to_string())
+            })?;
+            Ok((date - *epoch).num_milliseconds())
+        },
+    )?) as _;
+    Ok(ret)
+}
+
+#[inline]
+pub fn date64_interval_dt_op(
+    right: &Arc<dyn Array>,
+    epoch: &NaiveDate,
+    prior: &NaiveDate,
+    day_op: fn(NaiveDate, Days) -> Option<NaiveDate>,
+) -> Result<ArrayRef> {
+    let right: &PrimitiveArray<IntervalDayTimeType> = right.as_primitive();
+    let ret = Arc::new(try_unary::<IntervalDayTimeType, _, Date64Type>(
+        right,
+        |dt| {
+            let (days, millis) = IntervalDayTimeType::to_parts(dt);
+            let days_obj = Days::new(days.try_into().map_err(|_| {
+                DataFusionError::Internal(
+                    "Interval values cannot be casted as unsigned integers".to_string(),
+                )
+            })?);
+            let date = day_op(*prior, days_obj).ok_or_else(|| {
+                DataFusionError::Internal("Resulting date is out of range".to_string())
+            })?;
+            Ok((date - *epoch).num_milliseconds() - millis as i64)
+        },
+    )?) as _;
+    Ok(ret)
+}
+
+#[inline]
+pub fn date64_interval_mdn_op(
+    right: &Arc<dyn Array>,
+    epoch: &NaiveDate,
+    prior: &NaiveDate,
+    day_op: fn(NaiveDate, Days) -> Option<NaiveDate>,
+    month_op: fn(NaiveDate, Months) -> Option<NaiveDate>,
+) -> Result<ArrayRef> {
+    let cast_err = |_| {
+        DataFusionError::Internal(
+            "Interval values cannot be casted as unsigned integers".to_string(),
+        )
+    };
+    let out_of_range =
+        || DataFusionError::Internal("Resulting date is out of range".to_string());
+    let right: &PrimitiveArray<IntervalMonthDayNanoType> = right.as_primitive();
+    let ret = Arc::new(try_unary::<IntervalMonthDayNanoType, _, Date64Type>(
+        right,
+        |mdn| {
+            let (months, days, nanos) = IntervalMonthDayNanoType::to_parts(mdn);
+            let months_obj = Months::new(months.try_into().map_err(cast_err)?);
+            let month_diff = month_op(*prior, months_obj).ok_or_else(out_of_range)?;
+            let days_obj = Days::new(days.try_into().map_err(cast_err)?);
+            let value = day_op(month_diff, days_obj).ok_or_else(out_of_range)?;
+            Ok((value - *epoch).num_milliseconds() - nanos / 1_000_000)
+        },
+    )?) as _;
+    Ok(ret)
 }
 
 #[cfg(test)]
